@@ -1,33 +1,27 @@
 {
-  "_args": [
-    [
-      "babel-loader@8.2.2",
-      "/Users/chinmoy/1_WORK/15_Taxila/comproject"
-    ]
-  ],
-  "_development": true,
-  "_from": "babel-loader@8.2.2",
-  "_id": "babel-loader@8.2.2",
+  "_from": "babel-loader@^8.2.3",
+  "_id": "babel-loader@8.2.3",
   "_inBundle": false,
-  "_integrity": "sha512-JvTd0/D889PQBtUXJ2PXaKU/pjZDMtHA9V2ecm+eNRmmBCMR09a+fmpGTNwnJtFmFl5Ei7Vy47LjBb+L0wQ99g==",
+  "_integrity": "sha512-n4Zeta8NC3QAsuyiizu0GkmRcQ6clkV9WFUnUf1iXP//IeSKbWjofW3UHyZVwlOB4y039YQKefawyTn64Zwbuw==",
   "_location": "/babel-loader",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "babel-loader@8.2.2",
+    "raw": "babel-loader@^8.2.3",
     "name": "babel-loader",
     "escapedName": "babel-loader",
-    "rawSpec": "8.2.2",
+    "rawSpec": "^8.2.3",
     "saveSpec": null,
-    "fetchSpec": "8.2.2"
+    "fetchSpec": "^8.2.3"
   },
   "_requiredBy": [
     "/laravel-mix"
   ],
-  "_resolved": "https://registry.npmjs.org/babel-loader/-/babel-loader-8.2.2.tgz",
-  "_spec": "8.2.2",
-  "_where": "/Users/chinmoy/1_WORK/15_Taxila/comproject",
+  "_resolved": "https://registry.npmjs.org/babel-loader/-/babel-loader-8.2.3.tgz",
+  "_shasum": "8986b40f1a64cacfcb4b8429320085ef68b1342d",
+  "_spec": "babel-loader@^8.2.3",
+  "_where": "/Users/chinmoy/1_WORK/15_Taxila/Docker/comproject/src/node_modules/laravel-mix",
   "author": {
     "name": "Luis Couto",
     "email": "hello@luiscouto.pt"
@@ -47,12 +41,14 @@
   "bugs": {
     "url": "https://github.com/babel/babel-loader/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
     "find-cache-dir": "^3.3.1",
     "loader-utils": "^1.4.0",
     "make-dir": "^3.1.0",
     "schema-utils": "^2.6.5"
   },
+  "deprecated": false,
   "description": "babel module loader for webpack",
   "devDependencies": {
     "@ava/babel": "^1.0.1",
@@ -79,7 +75,7 @@
     "react-intl-webpack-plugin": "^0.3.0",
     "rimraf": "^3.0.0",
     "semver": "7.3.2",
-    "webpack": "^5.4.0"
+    "webpack": "^5.34.0"
   },
   "engines": {
     "node": ">= 8.9"
@@ -152,5 +148,5 @@
     "test": "yarn run lint && cross-env BABEL_ENV=test yarn run build && yarn run test-only",
     "test-only": "nyc ava"
   },
-  "version": "8.2.2"
+  "version": "8.2.3"
 }
