@@ -1,33 +1,27 @@
 {
-  "_args": [
-    [
-      "bytes@3.1.0",
-      "/Users/chinmoy/1_WORK/15_Taxila/comproject"
-    ]
-  ],
-  "_development": true,
-  "_from": "bytes@3.1.0",
-  "_id": "bytes@3.1.0",
+  "_from": "bytes@3.1.1",
+  "_id": "bytes@3.1.1",
   "_inBundle": false,
-  "_integrity": "sha512-zauLjrfCG+xvoyaqLoV8bLVXXNGC4JqlxFCutSDWA6fJrTo2ZuvLYTqZ7aHBLZSMOopbzwv8f+wZcVzfVTI2Dg==",
+  "_integrity": "sha512-dWe4nWO/ruEOY7HkUJ5gFt1DCFV9zPRoJr8pV0/ASQermOZjtq8jMjOprC0Kd10GLN+l7xaUPvxzJFWtxGu8Fg==",
   "_location": "/body-parser/bytes",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "bytes@3.1.0",
+    "raw": "bytes@3.1.1",
     "name": "bytes",
     "escapedName": "bytes",
-    "rawSpec": "3.1.0",
+    "rawSpec": "3.1.1",
     "saveSpec": null,
-    "fetchSpec": "3.1.0"
+    "fetchSpec": "3.1.1"
   },
   "_requiredBy": [
     "/body-parser"
   ],
-  "_resolved": "https://registry.npmjs.org/bytes/-/bytes-3.1.0.tgz",
-  "_spec": "3.1.0",
-  "_where": "/Users/chinmoy/1_WORK/15_Taxila/comproject",
+  "_resolved": "https://registry.npmjs.org/bytes/-/bytes-3.1.1.tgz",
+  "_shasum": "3f018291cb4cbad9accb6e6970bca9c8889e879a",
+  "_spec": "bytes@3.1.1",
+  "_where": "/Users/chinmoy/1_WORK/15_Taxila/Docker/comproject/src/node_modules/body-parser",
   "author": {
     "name": "TJ Holowaychuk",
     "email": "tj@vision-media.ca",
@@ -36,6 +30,7 @@
   "bugs": {
     "url": "https://github.com/visionmedia/bytes.js/issues"
   },
+  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Jed Watson",
@@ -46,11 +41,13 @@
       "email": "theo.fidry@gmail.com"
     }
   ],
+  "deprecated": false,
   "description": "Utility to parse a string bytes to bytes and vice-versa",
   "devDependencies": {
-    "eslint": "5.12.1",
-    "mocha": "5.2.0",
-    "nyc": "13.1.0"
+    "eslint": "7.32.0",
+    "eslint-plugin-markdown": "2.2.1",
+    "mocha": "9.1.3",
+    "nyc": "15.1.0"
   },
   "engines": {
     "node": ">= 0.8"
@@ -80,8 +77,8 @@
   "scripts": {
     "lint": "eslint .",
     "test": "mocha --check-leaks --reporter spec",
-    "test-ci": "nyc --reporter=text npm test",
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "3.1.0"
+  "version": "3.1.1"
 }
