@@ -1,33 +1,27 @@
 {
-  "_args": [
-    [
-      "css-declaration-sorter@6.1.1",
-      "/Users/chinmoy/1_WORK/15_Taxila/comproject"
-    ]
-  ],
-  "_development": true,
-  "_from": "css-declaration-sorter@6.1.1",
-  "_id": "css-declaration-sorter@6.1.1",
+  "_from": "css-declaration-sorter@^6.0.3",
+  "_id": "css-declaration-sorter@6.1.4",
   "_inBundle": false,
-  "_integrity": "sha512-BZ1aOuif2Sb7tQYY1GeCjG7F++8ggnwUkH5Ictw0mrdpqpEd+zWmcPdstnH2TItlb74FqR0DrVEieon221T/1Q==",
+  "_integrity": "sha512-lpfkqS0fctcmZotJGhnxkIyJWvBXgpyi2wsFd4J8VB7wzyrT6Ch/3Q+FMNJpjK4gu1+GN5khOnpU2ZVKrLbhCw==",
   "_location": "/css-declaration-sorter",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "css-declaration-sorter@6.1.1",
+    "raw": "css-declaration-sorter@^6.0.3",
     "name": "css-declaration-sorter",
     "escapedName": "css-declaration-sorter",
-    "rawSpec": "6.1.1",
+    "rawSpec": "^6.0.3",
     "saveSpec": null,
-    "fetchSpec": "6.1.1"
+    "fetchSpec": "^6.0.3"
   },
   "_requiredBy": [
     "/cssnano-preset-default"
   ],
-  "_resolved": "https://registry.npmjs.org/css-declaration-sorter/-/css-declaration-sorter-6.1.1.tgz",
-  "_spec": "6.1.1",
-  "_where": "/Users/chinmoy/1_WORK/15_Taxila/comproject",
+  "_resolved": "https://registry.npmjs.org/css-declaration-sorter/-/css-declaration-sorter-6.1.4.tgz",
+  "_shasum": "b9bfb4ed9a41f8dcca9bf7184d849ea94a8294b4",
+  "_spec": "css-declaration-sorter@^6.0.3",
+  "_where": "/Users/chinmoy/1_WORK/15_Taxila/Docker/comproject/src/node_modules/cssnano-preset-default",
   "author": {
     "name": "Selwyn",
     "email": "talk@selwyn.cc",
@@ -36,15 +30,17 @@
   "bugs": {
     "url": "https://github.com/Siilwyn/css-declaration-sorter/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
     "timsort": "^0.3.0"
   },
+  "deprecated": false,
   "description": "Sorts CSS declarations fast and automatically in a certain order.",
   "devDependencies": {
-    "@mdn/browser-compat-data": "^3.3.10",
-    "eslint": "^7.30.0",
-    "postcss": "^8.3.5",
-    "rollup": "^2.53.2",
+    "@mdn/browser-compat-data": "^3.3.14",
+    "eslint": "^7.32.0",
+    "postcss": "^8.3.8",
+    "rollup": "^2.57.0",
     "uvu": "^0.5.1"
   },
   "engines": {
@@ -92,5 +88,5 @@
     "test:ci": "npm test && npm run lint -- --max-warnings 0"
   },
   "type": "module",
-  "version": "6.1.1"
+  "version": "6.1.4"
 }
