@@ -1,33 +1,27 @@
 {
-  "_args": [
-    [
-      "css-select@4.1.3",
-      "/Users/chinmoy/1_WORK/15_Taxila/comproject"
-    ]
-  ],
-  "_development": true,
-  "_from": "css-select@4.1.3",
-  "_id": "css-select@4.1.3",
+  "_from": "css-select@^4.1.3",
+  "_id": "css-select@4.2.1",
   "_inBundle": false,
-  "_integrity": "sha512-gT3wBNd9Nj49rAbmtFHj1cljIAOLYSX1nZ8CB7TBO3INYckygm5B7LISU/szY//YmdiSLbJvDLOx9VnMVpMBxA==",
+  "_integrity": "sha512-/aUslKhzkTNCQUB2qTX84lVmfia9NyjP3WpDGtj/WxhwBzWBYUV3DgUpurHTme8UTPcPlAD1DJ+b0nN/t50zDQ==",
   "_location": "/css-select",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "css-select@4.1.3",
+    "raw": "css-select@^4.1.3",
     "name": "css-select",
     "escapedName": "css-select",
-    "rawSpec": "4.1.3",
+    "rawSpec": "^4.1.3",
     "saveSpec": null,
-    "fetchSpec": "4.1.3"
+    "fetchSpec": "^4.1.3"
   },
   "_requiredBy": [
     "/svgo"
   ],
-  "_resolved": "https://registry.npmjs.org/css-select/-/css-select-4.1.3.tgz",
-  "_spec": "4.1.3",
-  "_where": "/Users/chinmoy/1_WORK/15_Taxila/comproject",
+  "_resolved": "https://registry.npmjs.org/css-select/-/css-select-4.2.1.tgz",
+  "_shasum": "9e665d6ae4c7f9d65dbe69d0316e3221fb274cdd",
+  "_spec": "css-select@^4.1.3",
+  "_where": "/Users/chinmoy/1_WORK/15_Taxila/Docker/comproject/src/node_modules/svgo",
   "author": {
     "name": "Felix Boehm",
     "email": "me@feedic.com"
@@ -35,29 +29,30 @@
   "bugs": {
     "url": "https://github.com/fb55/css-select/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
     "boolbase": "^1.0.0",
-    "css-what": "^5.0.0",
-    "domhandler": "^4.2.0",
-    "domutils": "^2.6.0",
-    "nth-check": "^2.0.0"
+    "css-what": "^5.1.0",
+    "domhandler": "^4.3.0",
+    "domutils": "^2.8.0",
+    "nth-check": "^2.0.1"
   },
+  "deprecated": false,
   "description": "a CSS selector compiler/engine",
   "devDependencies": {
-    "@types/boolbase": "^1.0.0",
-    "@types/jest": "^26.0.14",
-    "@types/node": "^15.0.2",
-    "@typescript-eslint/eslint-plugin": "^4.1.0",
-    "@typescript-eslint/parser": "^4.1.0",
+    "@types/boolbase": "^1.0.1",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^17.0.4",
+    "@typescript-eslint/eslint-plugin": "^5.8.0",
+    "@typescript-eslint/parser": "^5.8.0",
     "cheerio-soupselect": "^0.1.1",
-    "coveralls": "^3.0.2",
-    "eslint": "^7.9.0",
+    "eslint": "^8.5.0",
     "eslint-config-prettier": "^8.1.0",
-    "htmlparser2": "^6.0.0",
-    "jest": "^27.0.3",
-    "prettier": "^2.1.2",
-    "ts-jest": "^27.0.1",
-    "typescript": "^4.0.2"
+    "htmlparser2": "^7.2.0",
+    "jest": "^27.4.5",
+    "prettier": "^2.5.1",
+    "ts-jest": "^27.1.2",
+    "typescript": "^4.5.4"
   },
   "files": [
     "lib"
@@ -91,7 +86,6 @@
   },
   "scripts": {
     "build": "tsc",
-    "coverage": "cat coverage/lcov.info | coveralls",
     "format": "npm run format:es && npm run format:prettier",
     "format:es": "npm run lint:es -- --fix",
     "format:prettier": "npm run prettier -- --write",
@@ -100,8 +94,9 @@
     "lint:prettier": "npm run prettier -- --check",
     "prepare": "npm run build",
     "prettier": "prettier '**/*.{ts,md,json,yml}'",
-    "test": "jest --coverage && npm run lint"
+    "test": "npm run test:jest && npm run lint",
+    "test:jest": "jest"
   },
   "types": "lib/index.d.ts",
-  "version": "4.1.3"
+  "version": "4.2.1"
 }
