@@ -1,33 +1,27 @@
 {
-  "_args": [
-    [
-      "css-what@5.0.1",
-      "/Users/chinmoy/1_WORK/15_Taxila/comproject"
-    ]
-  ],
-  "_development": true,
-  "_from": "css-what@5.0.1",
-  "_id": "css-what@5.0.1",
+  "_from": "css-what@^5.1.0",
+  "_id": "css-what@5.1.0",
   "_inBundle": false,
-  "_integrity": "sha512-FYDTSHb/7KXsWICVsxdmiExPjCfRC4qRFBdVwv7Ax9hMnvMmEjP9RfxTEZ3qPZGmADDn2vAKSo9UcN1jKVYscg==",
+  "_integrity": "sha512-arSMRWIIFY0hV8pIxZMEfmMI47Wj3R/aWpZDDxWYCPEiOMv6tfOrnpDtgxBYPEQD4V0Y/958+1TdC3iWTFcUPw==",
   "_location": "/css-what",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "css-what@5.0.1",
+    "raw": "css-what@^5.1.0",
     "name": "css-what",
     "escapedName": "css-what",
-    "rawSpec": "5.0.1",
+    "rawSpec": "^5.1.0",
     "saveSpec": null,
-    "fetchSpec": "5.0.1"
+    "fetchSpec": "^5.1.0"
   },
   "_requiredBy": [
     "/css-select"
   ],
-  "_resolved": "https://registry.npmjs.org/css-what/-/css-what-5.0.1.tgz",
-  "_spec": "5.0.1",
-  "_where": "/Users/chinmoy/1_WORK/15_Taxila/comproject",
+  "_resolved": "https://registry.npmjs.org/css-what/-/css-what-5.1.0.tgz",
+  "_shasum": "3f7b707aadf633baf62c2ceb8579b545bb40f7fe",
+  "_spec": "css-what@^5.1.0",
+  "_where": "/Users/chinmoy/1_WORK/15_Taxila/Docker/comproject/src/node_modules/css-select",
   "author": {
     "name": "Felix Böhm",
     "email": "me@feedic.com",
@@ -36,21 +30,22 @@
   "bugs": {
     "url": "https://github.com/fb55/css-what/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {},
+  "deprecated": false,
   "description": "a CSS selector parser",
   "devDependencies": {
-    "@types/jest": "^26.0.3",
-    "@types/node": "^15.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.1.0",
-    "@typescript-eslint/parser": "^4.1.0",
-    "coveralls": "^3.0.5",
-    "eslint": "^7.0.0",
+    "@types/jest": "^27.0.2",
+    "@types/node": "^16.10.3",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-node": "^11.1.0",
-    "jest": "^27.0.1",
-    "prettier": "^2.0.5",
-    "ts-jest": "^27.0.1",
-    "typescript": "^4.0.2"
+    "jest": "^27.2.4",
+    "prettier": "^2.4.1",
+    "ts-jest": "^27.0.5",
+    "typescript": "^4.4.3"
   },
   "engines": {
     "node": ">= 6"
@@ -77,7 +72,6 @@
   },
   "scripts": {
     "build": "tsc",
-    "coverage": "cat coverage/lcov.info | coveralls",
     "format": "npm run format:es && npm run format:prettier",
     "format:es": "npm run lint:es -- --fix",
     "format:prettier": "npm run prettier -- --write",
@@ -86,8 +80,9 @@
     "lint:prettier": "npm run prettier -- --check",
     "prepare": "npm run build",
     "prettier": "prettier '**/*.{ts,md,json,yml}'",
-    "test": "jest --coverage -u && npm run lint"
+    "test": "npm run test:jest && npm run lint",
+    "test:jest": "jest"
   },
   "types": "lib/index.d.ts",
-  "version": "5.0.1"
+  "version": "5.1.0"
 }
