@@ -1,35 +1,29 @@
 {
-  "_args": [
-    [
-      "domhandler@4.2.0",
-      "/Users/chinmoy/1_WORK/15_Taxila/comproject"
-    ]
-  ],
-  "_development": true,
-  "_from": "domhandler@4.2.0",
-  "_id": "domhandler@4.2.0",
+  "_from": "domhandler@^4.3.0",
+  "_id": "domhandler@4.3.0",
   "_inBundle": false,
-  "_integrity": "sha512-zk7sgt970kzPks2Bf+dwT/PLzghLnsivb9CcxkvR8Mzr66Olr0Ofd8neSbglHJHaHa2MadfoSdNlKYAaafmWfA==",
+  "_integrity": "sha512-fC0aXNQXqKSFTr2wDNZDhsEYjCiYsDWl3D01kwt25hm1YIPyDGHvvi3rw+PLqHAl/m71MaiF7d5zvBr0p5UB2g==",
   "_location": "/domhandler",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "domhandler@4.2.0",
+    "raw": "domhandler@^4.3.0",
     "name": "domhandler",
     "escapedName": "domhandler",
-    "rawSpec": "4.2.0",
+    "rawSpec": "^4.3.0",
     "saveSpec": null,
-    "fetchSpec": "4.2.0"
+    "fetchSpec": "^4.3.0"
   },
   "_requiredBy": [
     "/css-select",
     "/dom-serializer",
     "/domutils"
   ],
-  "_resolved": "https://registry.npmjs.org/domhandler/-/domhandler-4.2.0.tgz",
-  "_spec": "4.2.0",
-  "_where": "/Users/chinmoy/1_WORK/15_Taxila/comproject",
+  "_resolved": "https://registry.npmjs.org/domhandler/-/domhandler-4.3.0.tgz",
+  "_shasum": "16c658c626cf966967e306f966b431f77d4a5626",
+  "_spec": "domhandler@^4.3.0",
+  "_where": "/Users/chinmoy/1_WORK/15_Taxila/Docker/comproject/src/node_modules/css-select",
   "author": {
     "name": "Felix Boehm",
     "email": "me@feedic.com"
@@ -37,23 +31,24 @@
   "bugs": {
     "url": "https://github.com/fb55/domhandler/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
     "domelementtype": "^2.2.0"
   },
+  "deprecated": false,
   "description": "Handler for htmlparser2 that turns pages into a dom",
   "devDependencies": {
-    "@types/jest": "^26.0.0",
-    "@types/node": "^14.0.9",
-    "@typescript-eslint/eslint-plugin": "^4.1.0",
-    "@typescript-eslint/parser": "^4.1.0",
-    "coveralls": "^3.0.5",
-    "eslint": "^7.9.0",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^16.11.11",
+    "@typescript-eslint/eslint-plugin": "^5.5.0",
+    "@typescript-eslint/parser": "^5.5.0",
+    "eslint": "^8.3.0",
     "eslint-config-prettier": "^8.1.0",
-    "htmlparser2": "^6.0.0",
-    "jest": "^26.0.1",
-    "prettier": "^2.0.5",
-    "ts-jest": "^26.1.0",
-    "typescript": "^4.0.2"
+    "htmlparser2": "^7.2.0",
+    "jest": "^27.4.2",
+    "prettier": "^2.5.0",
+    "ts-jest": "^27.0.7",
+    "typescript": "^4.5.2"
   },
   "engines": {
     "node": ">= 4"
@@ -85,13 +80,13 @@
   },
   "scripts": {
     "build": "tsc",
-    "coverage": "cat coverage/lcov.info | coveralls",
     "format": "prettier --write '**/*.{ts,md,json}'",
     "lint": "eslint src",
     "prepare": "npm run build",
-    "test": "jest --coverage && npm run lint"
+    "test": "npm run test:jest && npm run lint",
+    "test:jest": "jest"
   },
   "sideEffects": false,
   "types": "lib/index.d.ts",
-  "version": "4.2.0"
+  "version": "4.3.0"
 }
